--- data/GY-91.csv
@@ -1,39 +1,0 @@
-Time,Acelerometer,Gyroscope,Magnetometer
-01:00:36,"[-0.001407, 0.002761, 0.994704]","[-0.172025, 0.201997, 0.055949]","[0.267079, 0.172684, 0.356691]"
-01:00:37,"[0.000302, 0.00154, 0.996658]","[0.072116, 0.41562, 0.208537]","[-0.445131, -1.208785, -1.070072]"
-01:00:38,"[-0.004581, 0.001785, 0.994704]","[-0.049954, 0.049409, -0.005086]","[0.445131, 0.0, -0.891727]"
-01:00:39,"[-0.001163, 0.00447, 0.996902]","[0.194186, -0.255766, 0.269572]","[0.267079, 0.863418, -0.713381]"
-01:00:40,"[-0.003116, 0.006423, 1.00032]","[-0.23306, -0.103178, -0.005086]","[0.089026, -1.726836, 0.178345]"
-01:00:41,"[-0.00336, 0.00032, 0.99739]","[0.194186, 0.263032, -0.218709]","[-1.335394, -0.690734, 0.891727]"
-01:00:42,"[-0.003848, 0.004226, 1.003738]","[0.163669, 0.232515, 0.116984]","[1.157341, -0.345367, -0.891727]"
-01:00:43,"[-0.001895, 0.004958, 0.994949]","[0.011081, -0.377837, 0.178019]","[0.801236, -0.345367, 0.891727]"
-01:00:44,"[-0.002139, 0.001052, 0.99739]","[0.011081, 0.049409, 0.178019]","[-0.445131, -0.172684, -0.713381]"
-01:00:45,"[-0.00336, 0.004226, 0.993972]","[-0.294095, 0.263032, 0.116984]","[-0.089026, 0.172684, 1.070072]"
-01:00:46,"[-0.003116, -0.000413, 0.996169]","[-0.141507, 0.018892, -0.066121]","[-0.089026, 0.172684, 0.713381]"
-01:00:47,"[-0.003116, 0.003738, 0.994949]","[0.072116, 0.140962, 0.055949]","[1.157341, -0.345367, 0.178345]"
-01:00:48,"[-0.005557, 0.005447, 1.004958]","[-0.019437, 0.079927, -0.127157]","[0.089026, -1.036102, -1.605108]"
-01:00:49,"[-0.007754, 0.00447, 0.998611]","[0.133151, -0.042143, -0.035604]","[0.623184, -0.518051, -1.070072]"
-01:00:50,"[-0.00336, -0.002122, 0.994949]","[-0.141507, -0.011626, 0.025431]","[0.801236, -1.036102, 0.891727]"
-01:00:51,"[-0.001163, 0.000808, 1.001785]","[-0.080472, 0.018892, 0.269572]","[0.267079, -0.863418, 0.0]"
-01:00:52,"[-0.004092, 0.001052, 0.997878]","[0.072116, -0.072661, -0.066121]","[0.623184, -0.172684, 0.356691]"
-01:00:53,"[0.001523, 0.004958, 0.999587]","[-0.507718, 0.140962, 0.116984]","[0.801236, -2.072203, -1.248417]"
-01:00:54,"[-0.003116, 0.006423, 0.993972]","[0.011081, 0.17148, -0.066121]","[-1.157341, 1.208785, -0.356691]"
-01:00:55,"[-0.001163, 0.00154, 1.000564]","[-0.019437, 0.17148, -0.096639]","[0.267079, -0.863418, 0.0]"
-01:00:56,"[-0.005069, 0.001296, 1.000808]","[0.163669, 0.385103, 0.086466]","[0.623184, -1.554152, 0.0]"
-01:00:57,"[-0.005801, 0.004714, 0.998855]","[0.133151, -0.072661, -0.035604]","[-0.267079, 0.0, 1.248417]"
-01:00:58,"[-0.005557, 0.003005, 0.995437]","[-0.23306, -0.225249, -0.035604]","[-0.979289, 0.0, 0.178345]"
-01:01:00,"[-0.001407, 0.00447, 0.993728]","[-0.11099, 0.17148, 0.116984]","[-0.801236, -1.89952, 0.0]"
-01:01:01,"[0.001279, 0.003982, 0.997878]","[-0.049954, 0.049409, 0.086466]","[0.267079, -0.518051, -0.356691]"
-01:01:02,"[-0.000918, 0.002761, 0.995437]","[-0.172025, 0.110445, -0.066121]","[-1.335394, -1.036102, -0.178345]"
-01:01:03,"[-0.004581, 0.005447, 0.996169]","[0.072116, 0.079927, -0.066121]","[-0.445131, -0.863418, 0.356691]"
-01:01:04,"[-0.005313, 0.003005, 0.995193]","[0.041598, -0.072661, 0.208537]","[0.267079, -1.208785, -0.713381]"
-01:01:05,"[-0.004581, -0.006272, 0.995437]","[0.194186, -0.042143, -0.249227]","[0.267079, 0.172684, 0.356691]"
-01:01:06,"[-0.004581, 0.003249, 0.995681]","[-0.11099, 0.232515, -0.005086]","[-0.445131, -1.554152, 0.0]"
-01:01:07,"[-0.004825, 0.003249, 0.994949]","[-0.23306, 0.232515, 0.086466]","[0.801236, 0.690734, -0.535036]"
-01:01:08,"[-0.005069, 0.005935, 0.995193]","[-0.049954, -0.011626, 0.025431]","[0.623184, -0.172684, 0.356691]"
-01:01:09,"[-0.000186, 0.005447, 0.998367]","[0.011081, -0.103178, -0.096639]","[1.691499, -0.863418, 0.713381]"
-01:01:10,"[-0.004092, 0.005691, 1.000076]","[0.133151, 0.17148, 0.178019]","[-0.089026, -0.172684, -1.070072]"
-01:01:11,"[-0.00336, 0.000808, 1.001052]","[-0.049954, -0.164214, 0.086466]","[0.089026, -0.690734, 0.535036]"
-01:01:12,"[-0.005557, 0.001052, 0.996413]","[0.072116, -0.133696, -0.005086]","[0.623184, -0.518051, -0.356691]"
-01:01:13,"[-0.000918, 0.003249, 0.996902]","[-0.141507, 0.17148, 0.147502]","[-1.691499, 0.0, -0.891727]"
-01:01:14,"[-0.006045, 0.00447, 0.995437]","[0.072116, 0.110445, 0.178019]","[-0.267079, -1.726836, -0.535036]"
