--- data/HDC-1080.csv
@@ -1,26 +1,0 @@
-Time,Temperature(C),Humidity
-01:02:29,33.94989013671875,24.40185546875
-01:02:30,33.97003173828125,24.40185546875
-01:02:32,33.97003173828125,24.40185546875
-01:02:33,33.9801025390625,24.40185546875
-01:02:34,33.97003173828125,24.40185546875
-01:02:36,33.97003173828125,24.40185546875
-01:02:37,33.97003173828125,24.40185546875
-01:02:38,33.97003173828125,24.40185546875
-01:02:39,33.97003173828125,24.40185546875
-01:02:41,33.97003173828125,24.40185546875
-01:02:42,33.9801025390625,24.40185546875
-01:02:43,33.97003173828125,24.40185546875
-01:02:45,33.9801025390625,24.40185546875
-01:02:46,33.9801025390625,24.40185546875
-01:02:47,33.9801025390625,24.40185546875
-01:02:48,33.97003173828125,24.40185546875
-01:02:50,33.9801025390625,24.40185546875
-01:02:51,33.9801025390625,24.40185546875
-01:02:52,33.9801025390625,24.40185546875
-01:02:54,33.9801025390625,24.40185546875
-01:02:55,34.000244140625,24.40185546875
-01:02:56,33.9801025390625,24.40185546875
-01:02:58,33.9801025390625,24.40185546875
-01:02:59,33.9801025390625,24.40185546875
-01:03:00,34.000244140625,24.40185546875
