--- data/NEO-6M.csv
@@ -1,1 +1,0 @@
-Time,Latitude,Longitude
