--- main_v2.py
@@ -1,15 +1,0 @@
-from modules import OctaSat
-
-OctaSat = OctaSat()
-
-if __name__ == '__main__':
-    while True:
-        try:
-            OctaSat.start()
-            
-        except OSError:
-            print('\n[ ! ] Warning: OSError, running anyways :).\n')
-
-        except KeyboardInterrupt:
-            print("\n[ ! ] Exiting\n")
-            exit()
