--- modules/HDC1080.py
@@ -1,20 +1,0 @@
-import sys
-from modules import SDL_Pi_HDC1080
-
-class HDC:
-	def __init__(self):
-		sys.path.append('./modules/SDL_Pi_HDC1080_Python3')
-		self.hdc1080 = SDL_Pi_HDC1080.SDL_Pi_HDC1080()
-
-	def temp(self, decimal):
-		temperature = round(self.hdc1080.readTemperature(), decimal)
-		return temperature
-
-	def hum(self, decimal):
-		humidity = round(self.hdc1080.readHumidity(), decimal)
-		return humidity
-
-# HDC = HDC()
-# print(HDC.hum(2))
-# print(HDC.temp(2))
-
